# swagger.yaml
openapi: 3.0.3
info:
  title: Menu Link API
  description: |
    API backend pour l'application SaaS de gestion de restaurant avec menu digital.
    
    ## Fonctionnalités
    - Multi-Restaurants (SaaS)
    - Gestion du compte Gérant et du Staff (Caissiers, Cuisine, Serveurs)
    - Prise de commande par QR code 
    - Séparation des vues (Client public vs Staff privé)
    
    ## Authentification
    Cette API utilise un système de jetons JWT. Un bouton "Authorize" se trouve ci-dessous pour inclure votre JWT dans les requêtes protégées. Les endpoints publics (ex: vue du menu depuis un QR code) ne nécessitent pas de token.
    
  version: 1.0.0
servers:
  - url: https://menulink-api.duckdns.org/api
    description: Production server (AWS EC2)
  - url: https://menu-link-api.onrender.com/api
    description: Backup server (Render)
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /categories:
    get:
      tags:
        - Categories
      summary: List all categories
      description: Récupère toutes les catégories, optionnellement filtrées par menu_id
      operationId: listCategories
      parameters:
        - name: menu_id
          in: query
          required: false
          description: Filtrer par ID du menu
          schema:
            type: string
            example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Liste des catégories
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
    
    post:
      tags:
        - Categories
      summary: Create a new category
      description: Crée une nouvelle catégorie
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - menu_id
                - name
              properties:
                menu_id:
                  type: string
                  example: "507f1f77bcf86cd799439011"
                name:
                  type: string
                  example: "Entrées"
                display_order:
                  type: integer
                  example: 0
      responses:
        '201':
          description: Catégorie créée
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Category'
                  message:
                    type: string
                    example: "Category created successfully"
  /categories/{id}:
    get:
      tags:
        - Categories
      summary: Get category by ID
      description: Récupère une catégorie spécifique
      operationId: getCategory
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la catégorie
          schema:
            type: string
            example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Catégorie trouvée
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Category'
        '404':
          description: Catégorie non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Categories
      summary: Update category
      description: Met à jour une catégorie existante
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la catégorie
          schema:
            type: string
            example: "507f1f77bcf86cd799439011"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Plats Principaux"
                display_order:
                  type: integer
                  example: 1
      responses:
        '200':
          description: Catégorie mise à jour
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Category'
                  message:
                    type: string
                    example: "Category updated successfully"
    
    delete:
      tags:
        - Categories
      summary: Delete category
      description: Supprime une catégorie (doit être vide)
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la catégorie
          schema:
            type: string
            example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Catégorie supprimée
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Category deleted successfully"
        '400':
          description: Impossible de supprimer (catégorie non vide)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /dishes:
    get:
      tags:
        - Dishes
      summary: List all dishes
      description: Récupère tous les plats, optionnellement filtrés par catégorie
      operationId: getDishes
      parameters:
        - name: category_id
          in: query
          required: false
          description: Filtrer par ID de catégorie
          schema:
            type: string
            example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Liste des plats
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dish'
    
    post:
      tags:
        - Dishes
      summary: Create a new dish
      description: Crée un nouveau plat
      operationId: createDish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
                - category_id
              properties:
                name:
                  type: string
                  example: "Ndolé"
                description:
                  type: string
                  example: "Feuilles de ndolé aux arachides avec viande"
                price:
                  type: number
                  format: float
                  example: 3000
                category_id:
                  type: string
                  example: "507f1f77bcf86cd799439011"
                image_url:
                  type: string
                  example: "https://example.com/images/ndole.jpg"
      responses:
        '201':
          description: Plat créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Dish'
                  message:
                    type: string
                    example: "Dish created successfully"
  /dishes/{id}:
    get:
      tags:
        - Dishes
      summary: Get dish details
      description: Récupère les détails d'un plat spécifique
      operationId: getDish
      parameters:
        - name: id
          in: path
          required: true
          description: ID du plat
          schema:
            type: string
            example: "507f1f77bcf86cd799439012"
      responses:
        '200':
          description: Plat trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Dish'
        '404':
          description: Plat non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Dishes
      summary: Update dish details
      description: Met à jour les détails d'un plat
      operationId: updateDish
      parameters:
        - name: id
          in: path
          required: true
          description: ID du plat
          schema:
            type: string
            example: "507f1f77bcf86cd799439012"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Poulet DG"
                description:
                  type: string
                  example: "Poulet Directeur Général avec plantains"
                price:
                  type: number
                  format: float
                  example: 4000
                is_available:
                  type: boolean
                  example: true
                image_url:
                  type: string
                  example: "https://example.com/images/poulet-dg.jpg"
      responses:
        '200':
          description: Plat mis à jour
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Dish'
                  message:
                    type: string
                    example: "Dish updated successfully"
    
    delete:
      tags:
        - Dishes
      summary: Delete a dish
      description: Supprime un plat
      operationId: deleteDish
      parameters:
        - name: id
          in: path
          required: true
          description: ID du plat
          schema:
            type: string
            example: "507f1f77bcf86cd799439012"
      responses:
        '200':
          description: Plat supprimé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Dish deleted successfully"
  /dishes/{id}/availability:
    patch:
      tags:
        - Dishes
      summary: Toggle dish availability
      description: Active/Désactive la disponibilité d'un plat (En stock / Rupture de stock)
      operationId: toggleAvailability
      parameters:
        - name: id
          in: path
          required: true
          description: ID du plat
          schema:
            type: string
            example: "507f1f77bcf86cd799439012"
      responses:
        '200':
          description: Disponibilité modifiée
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Dish'
                  message:
                    type: string
                    example: "Availability toggled successfully"
components:
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "ERROR_CODE"
            message:
              type: string
              example: "Error description"
            details:
              type: object
              additionalProperties: true
...
    Category:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        menu_id:
          type: string
          example: "507f1f77bcf86cd799439010"
        name:
          type: string
          example: "Entrées"
        display_order:
          type: integer
          example: 0
...
    Dish:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439012"
        name:
          type: string
          example: "Salade César"
        description:
          type: string
          example: "Salade fraîche avec croutons"
        price:
          type: number
          format: float
          example: 8.99
        is_available:
          type: boolean
          example: true
        image_url:
          type: string
          nullable: true
          example: "https://example.com/images/salade-cesar.jpg"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []